refactor: organize feature ui into mvvm packages
