+++ app/src/main/java/com/techmarketplace/MainActivity.kt
@@ -40,17 +40,17 @@ import androidx.core.content.getSystemService
 import com.techmarketplace.core.designsystem.TMTheme
 import com.techmarketplace.core.ui.BottomItem
 import com.techmarketplace.data.network.ApiClient
-import com.techmarketplace.data.repository.AuthRepository
-import com.techmarketplace.presentation.auth.view.LoginScreen
-import com.techmarketplace.presentation.auth.view.RegisterScreen
-import com.techmarketplace.presentation.auth.viewmodel.LoginViewModel
-import com.techmarketplace.presentation.cart.view.MyCartScreen
-import com.techmarketplace.presentation.home.view.AddProductRoute
-import com.techmarketplace.presentation.home.view.HomeRoute
-import com.techmarketplace.presentation.onboarding.view.WelcomeScreen
-import com.techmarketplace.presentation.orders.view.OrderScreen
-import com.techmarketplace.presentation.product.view.ProductDetailRoute
-import com.techmarketplace.presentation.profile.view.ProfileRoute
+import com.techmarketplace.feature.auth.data.AuthRepository
+import com.techmarketplace.feature.auth.presentation.login.LoginScreen
+import com.techmarketplace.feature.auth.presentation.login.LoginViewModel
+import com.techmarketplace.feature.auth.presentation.register.RegisterScreen
+import com.techmarketplace.feature.home.presentation.add.AddProductRoute
+import com.techmarketplace.feature.home.presentation.listings.HomeRoute
+import com.techmarketplace.feature.cart.presentation.MyCartScreen
+import com.techmarketplace.feature.onboarding.presentation.WelcomeScreen
+import com.techmarketplace.feature.orders.presentation.OrderScreen
+import com.techmarketplace.feature.product.presentation.detail.ProductDetailRoute
+import com.techmarketplace.feature.profile.presentation.ProfileRoute
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.location.Priority
 import kotlinx.coroutines.launch
