+++ app/src/main/java/com/techmarketplace/feature/auth/presentation/model/AuthFieldLimits.kt
@@ -1,0 +1,8 @@
+package com.techmarketplace.feature.auth.presentation.model
+
+internal object AuthFieldLimits {
+    const val MAX_NAME_LENGTH = 64
+    const val MAX_EMAIL_LENGTH = 254
+    const val MAX_PASSWORD_LENGTH = 128
+    const val MAX_CAMPUS_LENGTH = 64
+}
