+++ app/src/main/java/com/techmarketplace/feature/onboarding/presentation/WelcomeScreen.kt
@@ -1,0 +1,125 @@
+package com.techmarketplace.feature.onboarding.presentation
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.techmarketplace.R
+import com.techmarketplace.core.designsystem.Dimens
+import com.techmarketplace.core.designsystem.GreenDark
+import com.techmarketplace.core.designsystem.OffWhite
+
+@Composable
+fun WelcomeScreen(onContinue: () -> Unit) {
+    Surface(modifier = Modifier.fillMaxSize(), color = GreenDark) {
+        Column(modifier = Modifier.fillMaxSize()) {
+            Image(
+                painter = painterResource(R.drawable.welcome_photo),
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .fillMaxHeight(0.5f)
+                    .clip(RoundedCornerShape(bottomStart = 28.dp, bottomEnd = 28.dp))
+            )
+
+            Spacer(Modifier.height(12.dp))
+            OnboardingIndicator(total = 3, selectedIndex = 1, modifier = Modifier.align(Alignment.CenterHorizontally))
+            Spacer(Modifier.height(16.dp))
+
+            Box(Modifier.fillMaxWidth().weight(1f)) {
+                Column(
+                    modifier = Modifier
+                        .align(Alignment.Center)
+                        .offset(y = 20.dp)
+                        .padding(horizontal = 24.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text(
+                        text = stringResource(R.string.welcome_title),
+                        color = OffWhite,
+                        fontSize = 28.sp,
+                        lineHeight = 34.sp,
+                        fontWeight = FontWeight.SemiBold,
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier.fillMaxWidth()
+                    )
+                    Spacer(Modifier.height(12.dp))
+                    Text(
+                        text = stringResource(R.string.welcome_desc),
+                        color = OffWhite.copy(alpha = 0.7f),
+                        fontSize = 16.sp,
+                        lineHeight = 22.sp,
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier.fillMaxWidth()
+                    )
+                }
+            }
+
+            Button(
+                onClick = onContinue, // <-- navega
+                shape = RoundedCornerShape(28.dp),
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = OffWhite,
+                    contentColor = GreenDark
+                ),
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 24.dp, vertical = 24.dp)
+                    .height(Dimens.ButtonHeight)
+            ) {
+                Text("Continue", fontSize = 16.sp, fontWeight = FontWeight.SemiBold)
+            }
+            Spacer(Modifier.windowInsetsBottomHeight(WindowInsets.navigationBars))
+        }
+    }
+}
+
+@Composable
+private fun OnboardingIndicator(
+    total: Int,
+    selectedIndex: Int,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        repeat(total) { i ->
+            val selected = i == selectedIndex
+            if (selected) {
+                Box(
+                    modifier = Modifier
+                        .width(22.dp)
+                        .height(6.dp)
+                        .clip(RoundedCornerShape(50))
+                        .background(OffWhite)
+                )
+            } else {
+                Box(
+                    modifier = Modifier
+                        .size(6.dp)
+                        .clip(CircleShape)
+                        .background(OffWhite.copy(alpha = 0.6f))
+                )
+            }
+        }
+    }
+}
