--- app/src/main/java/com/techmarketplace/presentation/auth/model/AuthFieldLimits.kt
@@ -1,8 +1,0 @@
-package com.techmarketplace.presentation.auth.model
-
-internal object AuthFieldLimits {
-    const val MAX_NAME_LENGTH = 64
-    const val MAX_EMAIL_LENGTH = 254
-    const val MAX_PASSWORD_LENGTH = 128
-    const val MAX_CAMPUS_LENGTH = 64
-}
