--- app/src/main/java/com/techmarketplace/presentation/orders/view/OrdersScreen.kt
@@ -1,93 +1,0 @@
-package com.techmarketplace.presentation.orders.view
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.ArrowBack
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.techmarketplace.data.local.MyOrdersStore
-import java.text.NumberFormat
-import java.util.Locale
-import androidx.compose.material3.ExperimentalMaterial3Api
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun OrderScreen(
-    onNavigateBottom: (/* BottomItem u otro callback si lo necesitas */) -> Unit = {},
-    onBack: () -> Unit = {}
-) {
-    val orders by MyOrdersStore.orders.collectAsState()
-    val money = NumberFormat.getCurrencyInstance(Locale.getDefault())
-
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text("Orders") },
-                navigationIcon = {
-                    IconButton(onClick = onBack) {
-                        Icon(Icons.Outlined.ArrowBack, contentDescription = null)
-                    }
-                }
-            )
-        }
-    ) { inner ->
-        if (orders.isEmpty()) {
-            Column(
-                modifier = Modifier
-                    .padding(inner)
-                    .fillMaxSize()
-                    .padding(16.dp)
-            ) {
-                Text("No tienes órdenes aún.", style = MaterialTheme.typography.titleMedium)
-                Text(
-                    "Compra algo desde el detalle de un producto para verlas aquí.",
-                    style = MaterialTheme.typography.bodyMedium
-                )
-            }
-        } else {
-            LazyColumn(
-                modifier = Modifier
-                    .padding(inner)
-                    .fillMaxSize(),
-                contentPadding = PaddingValues(12.dp),
-                verticalArrangement = Arrangement.spacedBy(10.dp)
-            ) {
-                items(orders, key = { it.id }) { o ->
-                    Card(
-                        modifier = Modifier.fillMaxWidth(),
-                        colors = CardDefaults.cardColors()
-                    ) {
-                        Column(Modifier.padding(12.dp)) {
-                            Text("Order #${o.id.take(8)}", style = MaterialTheme.typography.titleMedium)
-                            Text("Listing: ${o.listingId}", style = MaterialTheme.typography.bodySmall)
-                            Text(
-                                "Total: " + money.format(o.totalCents / 100.0) + " ${o.currency}",
-                                style = MaterialTheme.typography.bodyMedium
-                            )
-                            Text("Status: ${o.status}", style = MaterialTheme.typography.bodySmall)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
